requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
lxml>=4.9.0
selenium>=4.15.0
python-dotenv>=1.0.0
webdriver-manager>=4.0.0
customtkinter>=5.2.0
pyinstaller>=6.0.0
Pillow>=10.0.0
openai>=1.0.0
anthropic>=0.18.0
google-genai>=0.2.0
tkinterweb>=3.0.0

